requirements.txt
boto3
python-telegram-bot
sentence-transformers
faiss-cpu
numpy